Altered BiomeList to use XML biome definitions file.
An external file is now used to define biomes, incl. their names,
colors, and temperature and humidity ranges. This file is currently
hardcoded to the name biomes_def.xml and is expected to be in the CWD.

The biome definitions also define a primary aspect for each biome. This
is currently nonfunctional, but will be used when the aspect system is
integrated.